+++ packages/@molt/command/src/types.ts
@@ -1,1 +1,5 @@
+import type { z } from 'zod'
+
 export type SchemaBase = 'ZodBoolean' | 'ZodNumber' | 'ZodString'
+
+export type SomeSchema = z.ZodRawShape
